test: move the full status test to newfeatures since old vtctld won't have this rpc
